backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sock==0.7.0
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-dotenv==1.0.1
